perf(handouts): lazy-load text version content
